[PATCH] getxattr error checking fix
From: Nathan Scott <nathans@sgi.com>,
      Andreas Gruenbacher <agruen@suse.de>

XFS regression tests tripped a couple of problems with the recent xattr fix.

When a size of 0 is passed in, the getxattr and listxattr syscalls return the
size that would be required for the value or list of names, without actually
returning the value.  The previous patch accidentally removed this test, and
so querying the required size broke.