[PATCH] fbdev state management
This adds some "state" information for power management to fbdev's,
along with a notifier mecanism to inform clients of state changes.  It
also "uses" this mecanism in the function fb_set_suspend() which was an
empty placeholder previously, and "shields" various places that access
the HW when state isn't running.  (It's best to not call them in the
first place, but the current state of fbcon makes that _very_ difficult)