[PATCH] Intel x86-64 support merge
This has all the x86-64 specific changes for Intel Prescott/Nocona
support.

It requires a few minor changes outside arch/x86_64, which I am sending
separately.

This patch is needed to boot an 64bit kernel on a 64-bit capable
Prescott machine.

The ugliest part is probably the swiotlb code.  In fact the code for
that is not even included, but just reused from IA64.  swiotlb
implements the PCI DMA API using bounce buffering.  I don't like this at
all, but there was no other way to support non DAC capable hardware
(like IDE or USB) on machines with >3GB.  Please redirect all flames for
that to the Intel chipset designers.

ChangeLog:
- Add Kconfig options for PSC
- Add support to reuse microcode driver from i386 (Suresh B Siddha)
- Try to optimize for the selected CPU
- Fix early CPUID check for Intel CPUs (Suresh B Siddha)
- Fix GDT to use the configured cache line size for padding
- Support monitor/mwait idle loop
- Support HyperThreading
- Support Intel CPUID flags
- Remove all 3dnow prefetches
- Add alternative() for the prefetchw prefetch inline.
- Include P4 driver in oprofile
- Support Intel NOPs in alternative