[PATCH] Add noinline attribute
From: Andi Kleen <ak@muc.de>

This patch adds the `noinline' function attribute.  It can be used to
explicitly tell the compiler to not inline functions.

We need this due to what is, IMO, a bug present in gcc-3.4 and current
gcc-3.5 CVS: the compiler is inlining init/main.c:rest_init() inside
init/main.c:start_kernel(), despite the fact that thay are declared to be
placed in different text sections.