[PATCH] ppc32: rework l2 cache code
This patch removes the code that tweaked the L1 cache when setting
up the L2 one. That was added a while ago in the intend of making
things more robust but ended up breaking earlier 750 CPU setup.

Also fix some crap in the L1 cache code that is only used for the
powermac sleep at this point.