[PATCH] PCI: Report meaningful error for failed resource allocation
pci_assign_resource reports odd messages when resource allocation fails.
This is because res->end and res->start are modified by allocate_resource.
For example:

PCI: Failed to allocate resource 1(0-ffffffff) for 0000:00:01.0

The following patch reports whether it's an IO or memory resource, and
includes the correct size.  For consistency, we report it in a similar
way to the failure message in pci_request_region(), even though
res->start is unlikely to be useful.