[PATCH] Re: 2.6.4-rc1 + hp100 EISA, not working
>>>>> "Pawel" == Pawe  <Pawe> writes:

Pawel> On 2.6.3 it just Oopsed and didn't work. On 2.6.4-rc1 it does
Pawel> not oops but it doesn't work (neither as module nor as build in kernel).

Note that hp100 in 2.6.4-rc1 is still broken wrt EISA probing, since
it lacks the terminating entry in the EISA ID list. This one-liner
should take care of random crashes that are often reported to LKML :

===== drivers/net/hp100.c 1.24 vs edited =====