[PATCH] Wireless pcmcia netdev patches
On Mon, Nov 17, 2003 at 02:57:44PM -0800, Jean Tourrilhes wrote:
> On Sun, Nov 16, 2003 at 11:33:01PM +0000, Russell King wrote:
> > Jean,
> >
> > Could you please arrange for this patch to be tested?  It covers all
> > the PCMCIA wireless drivers in 2.6.0-test9; I haven't been able to
> > test.
>
> 	Dual P500, kernel 2.6.0-test9 + your patch.
> -------------------------------------------------
>...
> 	Few comments :
> 	o Yanked the cards out of the socket, cardmgr did ifconfig
> down and removal of module automatically.
> 	o I did not see any patch for the Orinoco driver, but I did
> not check if it needs one. Did not look at HostAP either...
> 	o wl3501_cs : ask Arnaldo.
> 	o With 2.6.0-test9, I had instances of "can't read CIS". I
> actually had to reboot to perform the test.
> 	o The kobject Oops is something I think I saw before. I never
> remove my Aironet card because I don't want to reboot :-( This time,
> it seems to have survived better, so I see that as progress ;-)

This patch seems to have been forgotten about, so here it is again,
this time against 2.6.4-rc1.  This really needs to be send upstream,
though I don't know where we got to with this one.  Maybe we should
put it in akpm's tree for a while?

It's essentially more of the "don't deadlock if you pull out a
PCMCIA network card and it's currently configured up" patch which
went in recently for PCMCIA NIC cards.