[ARM PATCH] 1757/1: Allow building of LE kernels with BE-default toolchain
Patch from Deepak Saxena

The ARM makefile currently assumes that the compiler being used 
defaults to little-endian builds unless the big-endian options
are provided. This may not always be the case (my toolchains default
to BE since that's what I mostly work with) so this patch forces
endianess options to be passed to the toolchain. W/O this,
building without CONFIG_CPU_BIG_ENDIAN on a big-endian toolchain 
will still result in a big-endian kernel.